Added the 2 poison mushrooms
Poison mushrooms:

Did not connect the textures.
The transparency (black&white) is meant to be applied to both of the poison mushrooms.

The basket:
Is it possible to have the inside of the mushroom show the mesh? (it appears transparent because of inverted normals.

If not- the mesh will get altered.